+++ features/generator.feature
@@ -7,16 +7,8 @@ Feature: Generator
     Given The db is populated
     Given I am an existing non logged professor
     Then I log in
-    Then I go on class creation page
-    Then I create a class "foobar"
-    Then I create two students, "Jean" "Khule" and "Bill" "Boquet" for my class
-    Then I access the class test page
-    Then I click on add a test
-    Then I click on add a test online
-    Then I select the skill "T4-U5-A1b"
-    Then I add that competence
-    Then I add a name to that test
-    Then I create the test
+    Then I create the class "fuzzz", with students "Millui" "Cenlignes" and "Fatmodels" "Skinnycontrollers"
+    Then I create the test "fizz" for skill "T4-U5-A1b"
     Then I click on generate the question
     Given I am on the generator page
     Then I enter "0" as lower range and "60" as upper range
@@ -27,16 +19,8 @@ Feature: Generator
     Given The db is populated
     Given I am an existing non logged professor
     Then I log in
-    Then I go on class creation page
-    Then I create a class "foobar"
-    Then I create two students, "Jean" "Khule" and "Bill" "Boquet" for my class
-    Then I access the class test page
-    Then I click on add a test
-    Then I click on add a test online
-    Then I select the skill "T4-U5-A1b"
-    Then I add that competence
-    Then I add a name to that test
-    Then I create the test
+    Then I create the class "fuzzz", with students "Millui" "Cenlignes" and "Fatmodels" "Skinnycontrollers"
+    Then I create the test "fizz" for skill "T4-U5-A1b"
     Then I click on generate the question
     Given I am on the generator page
     When I click on the create button
@@ -46,16 +30,8 @@ Feature: Generator
     Given The db is populated
     Given I am an existing non logged professor
     Then I log in
-    Then I go on class creation page
-    Then I create a class "foobar"
-    Then I create two students, "Jean" "Khule" and "Bill" "Boquet" for my class
-    Then I access the class test page
-    Then I click on add a test
-    Then I click on add a test online
-    Then I select the skill "T4-U5-A1b"
-    Then I add that competence
-    Then I add a name to that test
-    Then I create the test
+    Then I create the class "fuzzz", with students "Millui" "Cenlignes" and "Fatmodels" "Skinnycontrollers"
+    Then I create the test "fizz" for skill "T4-U5-A1b"
     Then I click on generate the question
     Given I am on the generator page
     Then I enter "99999" as lower range and "-1" as upper range
@@ -66,16 +42,8 @@ Feature: Generator
     Given The db is populated
     Given I am an existing non logged professor
     Then I log in
-    Then I go on class creation page
-    Then I create a class "foobar"
-    Then I create two students, "Jean" "Khule" and "Bill" "Boquet" for my class
-    Then I access the class test page
-    Then I click on add a test
-    Then I click on add a test online
-    Then I select the skill "T4-U5-A1b"
-    Then I add that competence
-    Then I add a name to that test
-    Then I create the test
+    Then I create the class "fuzzz", with students "Millui" "Cenlignes" and "Fatmodels" "Skinnycontrollers"
+    Then I create the test "fizz" for skill "T4-U5-A1b"
     Then I click on generate the question
     Given I am on the generator page
     Then I enter "0" as lower range and "1" as upper range
