+++ features/steps/generator.py
@@ -35,55 +35,6 @@ def step_impl(context):
     assert br.find_element_by_id("id_error_panel")
 
 
-@then('I access the class test page')
-def step_impl(context):
-    br = context.browser
-    br.find_element_by_id("id_my_tests_button").click()
-    assert re.search(r'/professor/lesson/[0-9]+/test/', br.current_url)
-
-
-@then('I click on add a test')
-def step_impl(context):
-    br = context.browser
-    br.find_element_by_id("id_add_test_button").click()
-    assert re.search(r'/professor/lesson/[0-9]+/test/add', br.current_url)
-
-
-@then('I click on add a test online')
-def step_impl(context):
-    br = context.browser
-    br.find_element_by_id("id_add_test_online_button").click()
-    assert re.search(r'/professor/lesson/[0-9]+/test/online/add', br.current_url)
-
-
-@then('I select the skill "{skill_code}"')
-def step_impl(context, skill_code):
-    br = context.browser
-    Select(br.find_element_by_xpath("//select[@ng-model='stage13SelectedSkill']")).select_by_value(skill_code)
-
-
-@then('I add that competence')
-def step_impl(context):
-    br = context.browser
-    # TODO : This is actually really stupid, must adapt to a laggy response from Angular when clicking !
-    br.find_element_by_id("addSkillToTestButtonForStage13").click()
-    br.find_element_by_id("addSkillToTestButtonForStage13").click()
-    br.find_element_by_id("addSkillToTestButtonForStage13").click()
-
-
-@then('I add a name to that test')
-def step_impl(context):
-    br = context.browser
-    br.find_element_by_id("test_name").send_keys("foo_test")
-
-
-@then('I create the test')
-def step_impl(context):
-    br = context.browser
-    br.find_element_by_id("id_create_test_button").click()
-    # assert re.search(r'/professor/lesson/[0-9]+/test/online/[0-9]+/modify/', br.current_url)
-
-
 @then('I click on generate the question')
 def step_impl(context):
     br = context.browser
